FIX : Address deprecation warnings because of moved modules in chaco
This commit addresses deprecation warnings raised in chaco because of
the recent moves in chaco i.e. chaco.plots and chaco.overlays

	modified:   chaco/base_2d_plot.py
	modified:   chaco/base_xy_plot.py
	modified:   chaco/plot.py
	modified:   chaco/selectable_legend.py
	modified:   chaco/tests/test_array_or_none.py
	modified:   chaco/tools/image_inspector_tool.py
	modified:   chaco/tools/regression_lasso.py
