+++ chaco/tools/regression_lasso.py
@@ -21,8 +21,8 @@ from enable.api import ColorTrait, LineStyle
 from traits.api import Any, Float, Instance
 
 # Chaco imports
-from chaco.lasso_overlay import LassoOverlay
 from chaco.label import Label
+from chaco.overlays.lasso_overlay import LassoOverlay
 from .lasso_selection import LassoSelection
